[REFACTOR] Moved legacy text rendering method to TextDrawer as an optional feature
