+++ scan_creator.py
@@ -382,34 +382,6 @@ def _image_histogram(image: ImageType) -> ImageType: ...
 def _image_complexity(image: ImageType): ...
 
 
-def multiline_text_with_shade(
-    draw_obj: ImageDrawType, text: str,
-    pos: Tuple[int, int], offset: Tuple[int, int], spacing: int,
-    font: FreeTypeFont, text_color: Tuple[int, int, int], shade_color: Tuple[int, int, int]
-) -> None:
-    """
-    Draw multiline text with a shaded background on the image.
-
-    Args:
-        draw_obj (ImageDrawType): The ImageDraw object used to draw the text.
-        text (str): The text to be drawn.
-        pos (Tuple[int, int]): The starting position (x, y) for the text.
-        offset (Tuple[int, int]): The offset for drawing the shaded background behind the text.
-        spacing (int): The spacing between lines of text.
-        font (FreeTypeFont): The font to be used for drawing the text.
-        text_color (Tuple[int, int, int]): The color of the text.
-        shade_color (Tuple[int, int, int]): The color of the shaded background.
-
-    Returns:
-        None: This function does not return any value; it directly modifies the `draw_obj`.
-    """
-
-    x, y = pos
-    dx, dy = offset
-    draw_obj.multiline_text((x+dx, y+dy), text, fill=shade_color, font=font, spacing=spacing)
-    draw_obj.multiline_text((x, y), text, fill=text_color, font=font, spacing=spacing)
-
-    return None
 
 
 def create_scan_image(images: List[ImageType], grid: Tuple[int, int], snapshottimes: List[int], video_info: VideoInfo, logofile: str, use_text_drawer: bool) -> ImageType:
@@ -439,39 +411,7 @@ def create_scan_image(images: List[ImageType], grid: Tuple[int, int], snapshotti
         - Video information (size, duration, codec, etc.) is displayed at the top.
     """
 
-    # TODO: 分离读取、转换和验证步骤
-
-    def _parse_text_list(text_list: List[List[str | Dict[str, str]]], video_info: VideoInfo) -> List[List[str]]:
-        parsed_list: List[List[str]] = []
-        for row in text_list:
-            parsed_row: List[str] = []
-            for item in row:
-                if isinstance(item, dict) and "field" in item and "key" in item:
-                    parsed_row.append(video_info[item["field"]][item["key"]])
-                elif isinstance(item, str):
-                    parsed_row.append(item)
-                else:
-                    raise ValueError(f"Not support text_list item:{item}")
-            parsed_list.append(parsed_row)
-        return parsed_list
-
-    def _parse_font_list(font_idx_list: List[str], font_list: List[FreeTypeFont]) -> List[FreeTypeFont]:
-        parsed_list: List[FreeTypeFont] = []
-        for idx in font_idx_list:
-            if isinstance(idx, int):
-                parsed_list.append(font_list[idx])
-        return parsed_list
-
-    def _get_fonts(layout) -> List[FreeTypeFont]:
-        available_font_list: List[FreeTypeFont] = []
-        for font in layout["fonts"]:
-            if isinstance(font, dict) and "path" in font and "size" in font \
-                    and os.path.exists(font["path"]) and isinstance(font["size"], int):
-                available_font_list.append(ImageFont.truetype(font["path"], font["size"]))
-            else:
-                raise ValueError(f"{font} is not support")
-        return available_font_list
-
+    
     col, row = grid
     total_images = col * row
 
@@ -494,34 +434,11 @@ def create_scan_image(images: List[ImageType], grid: Tuple[int, int], snapshotti
 
     scan_image = Image.new("RGB", (canvas_width, canvas_height), "white")
     draw = ImageDraw.Draw(scan_image)
+    
+    text_drawer = TextDrawer(video_info=video_info, draw=draw, use_new_method=use_text_drawer)
+    text_drawer.draw_text()
 
-    with open("config/info_layout.json", mode='r', encoding='utf-8')as fp:
-        layout = json.load(fp)
-
-    spacing = layout["spacing"]
-    shade_offset = tuple(layout["shade_offset"])
-    text_color = tuple(layout["text_color"])
-    shade_color = tuple(layout["shade_color"])
-    pos_list = layout["pos_list"]
-    available_font_list: List[FreeTypeFont] = []
-    available_font_list = _get_fonts(layout)
-    time_font = available_font_list[layout["time_font"]]
-
-    # Parsing the `text_list` and setting it back to the config
-    text_list = _parse_text_list(layout["text_list"], video_info)
-
-    # 验证index
-    font_list = _parse_font_list(layout["font_list"], available_font_list)
-    if not use_text_drawer:
-        for i, j, k in zip(text_list, pos_list, font_list):
-            multiline_text_with_shade(draw, "\n".join(i), j, shade_offset, spacing, k, text_color, shade_color)
-            pass
-    else:
-        font_1 = font_list[0]
-        font_2 = font_list[1]
-        text_drawer = TextDrawer(video_info=video_info, draw=draw, font_1=font_1, font_2=font_2)
-        text_drawer.draw_text()
-
+    time_font = text_drawer.get_time_font()
     y_offset = 450
     for idx, image in enumerate(images):
 
@@ -638,7 +555,7 @@ def main():
             snapshots = take_snapshots(video_info, snapshot_times)
 
             scan = create_scan_image(snapshots, grid_shape, snapshot_times,
-                                     video_info, logo_file, False)
+                                     video_info, logo_file, True)
 
             w, h = scan.size
             scan = scan.resize((w//resize_scale, h//resize_scale), Resampling.LANCZOS)
